requirements.txt
absl-py==0.7.0
aiofiles==0.4.0
aiohttp==3.5.4
aniso8601==5.1.0
api==0.0.7
APScheduler==3.6.0
asn1crypto==0.24.0
astor==0.7.1
async-generator==1.10
async-timeout==3.0.1
attrs==18.1.0
Automat==0.6.0
bert-serving-client==1.8.9
boto==2.49.0
boto3==1.9.114
botocore==1.12.114
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
Click==7.0
cloudpickle==0.6.1
colorama==0.4.1
colorclass==2.2.0
coloredlogs==10.0
colorhash==1.0.2
ConfigArgParse==0.13.0
constantly==15.1.0
cryptography==2.2.2
cssselect==1.0.3
cycler==0.10.0
decorator==4.4.0
docopt==0.6.2
docutils==0.14
dukpy==0.2.0
fakeredis==0.10.3
fbmessenger==5.6.0
Flask==1.0.2
Flask-APScheduler==1.11.0
Flask-Cors==3.0.7
Flask-JWT-Simple==0.0.3
Flask-RESTful==0.3.7
freeze==1.0.10
future==0.17.1
gast==0.2.2
gensim==3.8.0
get==2018.11.19
gevent==1.4.0
GPUtil==1.4.0
greenlet==0.4.15
grpcio==1.19.0
h5py==2.9.0
humanfriendly==4.18
hyperlink==18.0.0
idna==2.6
idna-ssl==1.1.0
incremental==17.5.0
itsdangerous==1.1.0
javascripthon==0.10
jieba==0.39
Jinja2==2.10
jmespath==0.9.4
joblib==0.13.2
JPype1==0.6.3
JsonForm==0.0.2
jsonpickle==0.9.6
jsonschema==2.6.0
JsonSir==0.0.2
jupyter-echarts-pypkg==0.1.1
kafka-python==1.4.6
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
klein==17.10.0
lml==0.0.2
lxml==4.2.1
macropy3==1.1.0b2
Markdown==3.0.1
MarkupSafe==1.0
matplotlib==2.2.4
mattermostwrapper==2.1
mock==2.0.0
multidict==4.5.2
networkx==2.3
nltk==3.2.5
nose==1.3.7
numpy==1.14.5
overrides==1.9
packaging==17.1
paho-mqtt==1.4.0
pandas==0.23.0
parsel==1.4.0
pathlib==1.0.1
pbr==5.1.3
pika==0.11.2
post==2018.11.20
prompt-toolkit==1.0.14
protobuf==3.7.0
public==2018.11.20
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.18
PyDispatcher==2.0.5
pydot==1.4.1
pyecharts==0.5.5
pyecharts-javascripthon==0.0.6
pyecharts-jupyter-installer==0.0.3
Pygments==2.3.1
PyHamcrest==1.9.0
PyInquirer==1.0.3
PyJWT==1.7.1
pykwalify==1.6.0
pyltp==0.2.1
pymongo==3.8.0
PyMySQL==0.8.0
pyOpenSSL==18.0.0
pyparsing==2.3.1
PyQt5==5.10.1
pyqt5-tools==5.9.0.1.2
pyreadline==2.1
pyrsistent==0.14.11
PySocks==1.6.8
python-crfsuite==0.9.6
python-dateutil==2.8.0
Python-EasyConfig==0.1.7
python-engineio==3.8.2.post1
python-socketio==2.1.2
python-telegram-bot==10.1.0
pytz==2018.9
pywin32==224
PyYAML==3.13
pyzmq==18.0.2
query-string==2018.11.20
queuelib==1.5.0
rasa-core==0.12.3
rasa-core-sdk==0.12.1
rasa-nlu==0.13.4
rasa-nlu-gao==0.2.7
rasa-sdk==1.1.0
redis==2.10.6
regex==2019.3.12
relativetimebuilder==0.2.0
request==2018.11.20
requests==2.22.0
requests-toolbelt==0.9.1
Resource==0.2.1
RESTfulEf==0.1.1
rocketchat-API==0.6.27
ruamel.yaml==0.15.89
s3transfer==0.2.0
sanic-jwt==1.3.2
scikit-learn==0.19.0
scipy==1.2.1
Scrapy==1.5.0
service-identity==17.0.0
simplejson==3.16.0
sip==4.19.8
six==1.12.0
sklearn-crfsuite==0.3.6
slackclient==1.3.1
smart-open==1.8.4
sortedcontainers==2.1.0
SQLAlchemy==1.3.5
tabulate==0.8.3
tensorboard==1.10.0
tensorflow==1.10.0
termcolor==1.1.0
terminaltables==3.1.0
tqdm==4.31.1
twilio==6.25.2
twisted==18.9.0
typing==3.6.6
typing-extensions==3.7.4
tzlocal==1.5.1
urllib3==1.24.1
virtualenv==16.4.3
w3lib==1.19.0
wcwidth==0.1.7
webexteamssdk==1.1.1
websocket-client==0.54.0
websockets==6.0
Werkzeug==0.14.1
xlrd==1.1.0
yarl==1.3.0
zope.interface==4.5.0
